requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.0
openai==1.12.0
googletrans==3.1.0a0
